make the bkgnd surface and cursor as optional
cursor is disabled by default, if cursor-enable flag is set as true
in weston.ini, then only it will be created. In this case a background
surface also will be created with the surf-id 1000000 even it is not
configured in weston.ini, otherwise the cursor can't be set without a
focused surface.

Signed-off-by: Maniraj Devadoss <Maniraj.Devadoss@in.bosch.com>